Estatísticas

+++ src/main/java/net/ism/crud/clientsimple/controller/ClienteController.java
@@ -1,6 +1,7 @@
 package net.ism.crud.clientsimple.controller;
 
 import java.net.URI;
+import java.time.LocalDate;
 import java.util.List;
 
 import javax.servlet.http.HttpServletRequest;
@@ -38,7 +39,9 @@ public class ClienteController {
 	
 	@GetMapping
 	public ResponseEntity<List<ClienteMinimoDTO>> listar() {
+		// Lista clientes
 		List<Cliente> clientes = clienteRepo.findAll();
+		// Retorno
 		return ResponseEntity.ok(ClienteMinimoDTO.toList(clientes));
 	}
 	
@@ -46,17 +49,22 @@ public class ClienteController {
 	public ResponseEntity<ClienteDetalheDTO> detalhar(@PathVariable Long id) {
 		// Valida
 		if(!existsCliente(id)) return ResponseEntity.notFound().build();
-		
+		// Busca cliente
 		Cliente cliente = clienteRepo.getOne(id);
+		// Retorno
 		return ResponseEntity.ok(new ClienteDetalheDTO(cliente));
 	}
 	
 	@PostMapping @Transactional
 	public ResponseEntity<ClienteDetalheDTO> inserir(@RequestBody @Valid ClienteInsertFormDTO input, UriComponentsBuilder uriBuilder, HttpServletRequest request) {
+		// Data
+		LocalDate agora = LocalDate.now();
+		// Insere cliente
 		Cliente cliente = input.toModel();
 		clienteRepo.save(cliente);
-		estatisticaController.inserir(request.getRemoteAddr(), cliente);
-		
+		// Insere estatistica
+		estatisticaController.inserir(request.getRemoteAddr(), agora, cliente);
+		// Retorno
 		URI uri = uriBuilder.path("/clientes/{id}").buildAndExpand(cliente.getId()).toUri();
 		return ResponseEntity.created(uri).body(new ClienteDetalheDTO(cliente));
 	}
@@ -64,17 +72,21 @@ public class ClienteController {
 	public ResponseEntity<ClienteDetalheDTO> atualizar(@PathVariable Long id, @RequestBody @Valid ClienteUpdateFormDTO input) {
 		// Valida
 		if(!existsCliente(id)) return ResponseEntity.notFound().build();
-				
+		// Atualiza
 		Cliente cliente = input.update(id, clienteRepo);
+		// Retorno
 		return ResponseEntity.ok(new ClienteDetalheDTO(cliente));
 	}
 	@DeleteMapping("/{id}") @Transactional
 	public ResponseEntity<?> remover(@PathVariable Long id) {
 		// Valida
 		if(!existsCliente(id)) return ResponseEntity.notFound().build();
-				
-		estatisticaController.remover(id);
+		// Remove estatistica
+		boolean result = estatisticaController.remover(id);
+		if(result == EstatisticaController.OPERATION_FAIL) return ResponseEntity.status(500).build();
+		// Remove cliente
 		clienteRepo.deleteById(id);
+		// Retorno
 		return ResponseEntity.ok().build();
 	}
